requirements.txt
Django==5.2.6
django-cors-headers==4.2.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
psycopg2-binary
python-dotenv==1.1.0
